Move parallel workspace listing to the exporter implementation
